pyproject.toml
[project]
name = "plataforma-bd2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.7",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "watchdog>=6.0.0",
]
